Created a 'draw_line" function
The draw_line function will draw a horizontal or vertical line on the terminal emulator. This will be used to draw volume bar looking objects for the attack, decay, sustain and release controls of the ADSR Envelope. The potentiometers moving up and down will be visually represented on the terminal emulator.